feat: Separate rooms based on IP

+++ packages/webrtc/server/index.ts
@@ -10,7 +10,7 @@ import {
   SocketEventParams,
 } from "../types/signaling";
 import { CONNECTION_STATE, Member, ServerSocket } from "../types/server";
-import { getLocalIp } from "./utils";
+import { getIpByRequest, getLocalIp } from "./utils";
 
 const app = express();
 app.use(express.static("build/static"));
@@ -19,22 +19,33 @@ const io = new Server<ClientHandler, ServerHandler>(httpServer);
 
 const authenticate = new WeakMap<ServerSocket, string>();
 const mapper = new Map<string, Member>();
+const rooms = new Map<string, string[]>();
 
 io.on("connection", socket => {
   socket.on(CLINT_EVENT.JOIN_ROOM, ({ id, device }) => {
+    // 验证
     if (!id) return void 0;
     authenticate.set(socket, id);
+    // 加入房间
+    const ip = getIpByRequest(socket.request);
+    const room = rooms.get(ip) || [];
+    rooms.set(ip, [...room, id]);
+    mapper.set(id, { socket, device, state: CONNECTION_STATE.NORMAL, ip });
+    // 房间通知消息
     const initialization: SocketEventParams["JOINED_MEMBER"]["initialization"] = [];
-    mapper.forEach((value, key) => {
-      initialization.push({ id: key, device: value.device });
-      value.socket.emit(SERVER_EVENT.JOINED_ROOM, { id, device });
+    room.forEach(key => {
+      const instance = mapper.get(key);
+      if (!instance) return void 0;
+      initialization.push({ id: key, device: instance.device });
+      instance.socket.emit(SERVER_EVENT.JOINED_ROOM, { id, device });
     });
-    mapper.set(id, { socket, device, state: CONNECTION_STATE.NORMAL });
     socket.emit(SERVER_EVENT.JOINED_MEMBER, { initialization });
   });
 
   socket.on(CLINT_EVENT.SEND_OFFER, ({ origin, sdp, target }) => {
+    // 验证
     if (authenticate.get(socket) !== origin) return void 0;
+    // 转发`Offer`
     const targetSocket = mapper.get(target)?.socket;
     if (targetSocket) {
       targetSocket.emit(SERVER_EVENT.FORWARD_OFFER, { origin, sdp, target });
@@ -42,7 +53,9 @@ io.on("connection", socket => {
   });
 
   socket.on(CLINT_EVENT.SEND_ANSWER, ({ origin, sdp, target }) => {
+    // 验证
     if (authenticate.get(socket) !== origin) return void 0;
+    // 转发`Answer`
     const targetSocket = mapper.get(target)?.socket;
     if (targetSocket) {
       targetSocket.emit(SERVER_EVENT.FORWARD_ANSWER, { origin, sdp, target });
@@ -50,19 +63,45 @@ io.on("connection", socket => {
   });
 
   socket.on(CLINT_EVENT.LEAVE_ROOM, ({ id }) => {
+    // 验证
     if (authenticate.get(socket) !== id) return void 0;
+    // 退出房间
+    const instance = mapper.get(id);
+    if (!instance) return void 0;
+    const room = (rooms.get(instance.ip) || []).filter(key => key !== id);
+    if (room.length === 0) {
+      rooms.delete(instance.ip);
+    } else {
+      rooms.set(instance.ip, room);
+    }
     mapper.delete(id);
-    mapper.forEach(value => {
-      value.socket.emit(SERVER_EVENT.LEFT_ROOM, { id });
+    // 房间内通知
+    room.forEach(key => {
+      const instance = mapper.get(key);
+      if (!instance) return void 0;
+      instance.socket.emit(SERVER_EVENT.LEFT_ROOM, { id });
     });
   });
 
   socket.on("disconnect", () => {
+    // 验证
     const id = authenticate.get(socket);
+    // 退出房间
     if (id) {
+      const instance = mapper.get(id);
+      if (!instance) return void 0;
+      const room = (rooms.get(instance.ip) || []).filter(key => key !== id);
+      if (room.length === 0) {
+        rooms.delete(instance.ip);
+      } else {
+        rooms.set(instance.ip, room);
+      }
       mapper.delete(id);
-      mapper.forEach(value => {
-        value.socket.emit(SERVER_EVENT.LEFT_ROOM, { id });
+      // 房间内通知
+      room.forEach(key => {
+        const instance = mapper.get(key);
+        if (!instance) return void 0;
+        instance.socket.emit(SERVER_EVENT.LEFT_ROOM, { id });
       });
     }
   });
